# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for The Front Porch Bulletin
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Photo {
  url: string;
  credit: string;
}

export interface Article {
  id: number;
  title: string;
  subtitle?: string | null;
  content: string;
  author: string;
  category: string;
  featured: boolean;
  archived: boolean;
  photoUrl?: string | null;
  photos?: Photo[] | null;
  publishedAt: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateArticleBody {
  title: string;
  subtitle?: string | null;
  content: string;
  author: string;
  category: string;
  featured?: boolean;
  photoUrl?: string | null;
  photos?: Photo[] | null;
  publishedAt?: string;
}

export interface UpdateArticleBody {
  title?: string;
  subtitle?: string | null;
  content?: string;
  author?: string;
  category?: string;
  featured?: boolean;
  photoUrl?: string | null;
  photos?: Photo[] | null;
  publishedAt?: string;
}

export type ArticlesSummaryByCategoryItem = {
  category: string;
  count: number;
};

export interface ArticlesSummary {
  totalArticles: number;
  totalCategories: number;
  byCategory: ArticlesSummaryByCategoryItem[];
  recentCount: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  articleCount: number;
}

export interface AdminUser {
  clerkUserId: string;
  role: string;
  grantedAt: string;
  email?: string | null;
  name?: string | null;
}

export type SetUserRoleBodyRole =
  (typeof SetUserRoleBodyRole)[keyof typeof SetUserRoleBodyRole];

export const SetUserRoleBodyRole = {
  admin: "admin",
  approved_user: "approved_user",
} as const;

export interface SetUserRoleBody {
  role: SetUserRoleBodyRole;
}

export interface MyRoleResponse {
  clerkUserId: string;
  role?: string | null;
  isAdmin: boolean;
  isApproved: boolean;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface Obituary {
  id: number;
  name: string;
  birthDate?: string | null;
  deathDate?: string | null;
  hometown?: string | null;
  content: string;
  photoUrl?: string | null;
  photos?: Photo[] | null;
  createdAt: string;
  updatedAt: string;
}

export interface ObituaryBody {
  name: string;
  birthDate?: string | null;
  deathDate?: string | null;
  hometown?: string | null;
  content: string;
  photoUrl?: string | null;
  photos?: Photo[] | null;
}

export interface Spotlight {
  id: number;
  name: string;
  school: string;
  grade: string;
  description: string;
  photoUrl?: string | null;
  photos?: Photo[] | null;
  createdAt: string;
  updatedAt: string;
}

export interface SpotlightBody {
  name: string;
  school: string;
  grade: string;
  description: string;
  photoUrl?: string | null;
  photos?: Photo[] | null;
}

export interface BusinessSpotlight {
  id: number;
  name: string;
  businessType: string;
  description: string;
  photoUrl?: string | null;
  photos?: Photo[] | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface BusinessSpotlightBody {
  name: string;
  businessType: string;
  description: string;
  photoUrl?: string | null;
  photos?: Photo[] | null;
}

export interface GroupSpotlight {
  id: number;
  name: string;
  groupType: string;
  description: string;
  photoUrl?: string | null;
  photos?: Photo[] | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface GroupSpotlightBody {
  name: string;
  groupType: string;
  description: string;
  photoUrl?: string | null;
  photos?: Photo[] | null;
}

export interface Church {
  id: number;
  name: string;
  address: string;
  pastor: string;
  serviceTimes: string;
  phone: string;
  sortOrder: number;
  photoUrl?: string | null;
  photos?: Photo[] | null;
  createdAt: string;
}

export interface ChurchBody {
  name: string;
  address: string;
  pastor: string;
  serviceTimes: string;
  phone: string;
  sortOrder?: number;
  photoUrl?: string | null;
  photos?: Photo[] | null;
}

export type ListArticlesParams = {
  category?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
  includeArchived?: boolean;
};

export type ListArticles200 = {
  articles: Article[];
  total: number;
};

export type GetFeaturedArticles200 = {
  headline?: Article;
  secondary: Article[];
};

export type ArchiveArticleBody = {
  archived: boolean;
};

export type ListObituaries200 = {
  obituaries: Obituary[];
};

export type ListChurches200 = {
  churches: Church[];
};

export type ListCategories200 = {
  categories: Category[];
};

export type ListAdminUsers200 = {
  users: AdminUser[];
};