# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for The Front Porch Bulletin
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List all articles
      parameters:
        - name: category
          in: query
          schema:
            type: string
          required: false
        - name: featured
          in: query
          schema:
            type: boolean
          required: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          required: false
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          required: false
        - name: includeArchived
          in: query
          schema:
            type: boolean
            default: false
          required: false
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Article"
                  total:
                    type: integer
                required:
                  - articles
                  - total
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create a new article (requires auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "201":
          description: Article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
...
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get a single article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
    put:
      operationId: updateArticle
      tags: [articles]
      summary: Update an article (requires auth)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleBody"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete an article (requires auth)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
  /articles/{id}/archive:
    patch:
      operationId: archiveArticle
      tags: [articles]
      summary: Toggle archived status of an article (requires auth)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                archived:
                  type: boolean
              required:
                - archived
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          description: Unauthorized
...
  /business-spotlight:
    get:
      operationId: getBusinessSpotlight
      tags: [spotlight]
      summary: Get the current business spotlight
      responses:
        "200":
          description: Current business spotlight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessSpotlight"
        "404":
          description: No business spotlight set
    put:
      operationId: updateBusinessSpotlight
      tags: [spotlight]
      summary: Replace the business spotlight (requires auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessSpotlightBody"
      responses:
        "200":
          description: Updated business spotlight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessSpotlight"
        "401":
          description: Unauthorized
...
  /spotlight:
    get:
      operationId: getSpotlight
      tags: [spotlight]
      summary: Get the current student spotlight
      responses:
        "200":
          description: Current spotlight or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spotlight"
        "404":
          description: No spotlight set
    put:
      operationId: updateSpotlight
      tags: [spotlight]
      summary: Replace the student spotlight (requires auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpotlightBody"
      responses:
        "200":
          description: Updated spotlight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spotlight"
        "401":
          description: Unauthorized
  /churches:
    get:
      operationId: listChurches
      tags: [churches]
      summary: List all churches
      responses:
        "200":
          description: List of churches
          content:
            application/json:
              schema:
                type: object
                properties:
                  churches:
                    type: array
                    items:
                      $ref: "#/components/schemas/Church"
                required:
                  - churches
    post:
      operationId: createChurch
      tags: [churches]
      summary: Add a church (requires auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChurchBody"
      responses:
        "201":
          description: Created church
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Church"
        "401":
          description: Unauthorized
  /churches/{id}:
    put:
      operationId: updateChurch
      tags: [churches]
      summary: Update a church (requires auth)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChurchBody"
      responses:
        "200":
          description: Updated church
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Church"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteChurch
      tags: [churches]
      summary: Delete a church (requires auth)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    Photo:
      type: object
      properties:
        url:
          type: string
        credit:
          type: string
      required:
        - url
        - credit
    Article:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        content:
          type: string
        author:
          type: string
        category:
          type: string
        featured:
          type: boolean
        archived:
          type: boolean
        photoUrl:
          type: string
          nullable: true
        photos:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Photo"
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - author
        - category
        - featured
        - archived
        - publishedAt
        - createdAt
        - updatedAt
    CreateArticleBody:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        content:
          type: string
        author:
          type: string
        category:
          type: string
        featured:
          type: boolean
          default: false
        photoUrl:
          type: string
          nullable: true
        photos:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Photo"
        publishedAt:
          type: string
          format: date-time
      required:
        - title
        - content
        - author
        - category
    UpdateArticleBody:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        content:
          type: string
        author:
          type: string
        category:
          type: string
        featured:
          type: boolean
        photoUrl:
          type: string
          nullable: true
        photos:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Photo"
        publishedAt:
          type: string
          format: date-time
...
    Spotlight:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        school:
          type: string
        grade:
          type: string
        description:
          type: string
        photoUrl:
          type: string
          nullable: true
        photos:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Photo"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - school
        - grade
        - description
        - createdAt
        - updatedAt
    SpotlightBody:
      type: object
      properties:
        name:
          type: string
        school:
          type: string
        grade:
          type: string
        description:
          type: string
        photoUrl:
          type: string
          nullable: true
        photos:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Photo"
      required:
        - name
        - school
        - grade
        - description
    BusinessSpotlight:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        businessType:
          type: string
        description:
          type: string
        photoUrl:
          type: string
          nullable: true
        photos:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Photo"
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - businessType
        - description
        - status
        - createdAt
        - updatedAt
    BusinessSpotlightBody:
      type: object
      properties:
        name:
          type: string
        businessType:
          type: string
        description:
          type: string
        photoUrl:
          type: string
          nullable: true
        photos:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Photo"
      required:
        - name
        - businessType
        - description
...
    Church:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        pastor:
          type: string
        serviceTimes:
          type: string
        phone:
          type: string
        sortOrder:
          type: integer
        photoUrl:
          type: string
          nullable: true
        photos:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Photo"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - address
        - pastor
        - serviceTimes
        - phone
        - sortOrder
        - createdAt
    ChurchBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        pastor:
          type: string
        serviceTimes:
          type: string
        phone:
          type: string
        sortOrder:
          type: integer
        photoUrl:
          type: string
          nullable: true
        photos:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Photo"
      required:
        - name
        - address
        - pastor
        - serviceTimes
        - phone