# lib/api-zod/src/generated/types/churchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for The Front Porch Bulletin
 * OpenAPI spec version: 0.1.0
 */
import type { Photo } from "./photo";

export interface ChurchBody {
  name: string;
  address: string;
  pastor: string;
  serviceTimes: string;
  phone: string;
  sortOrder?: number;
  photoUrl?: string | null;
  photos?: Photo[] | null;
}