# lib/api-zod/src/generated/types/photo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for The Front Porch Bulletin
 * OpenAPI spec version: 0.1.0
 */

export interface Photo {
  url: string;
  credit: string;
}