# lib/api-zod/src/generated/types/spotlight.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for The Front Porch Bulletin
 * OpenAPI spec version: 0.1.0
 */
import type { Photo } from "./photo";

export interface Spotlight {
  id: number;
  name: string;
  school: string;
  grade: string;
  description: string;
  photoUrl?: string | null;
  photos?: Photo[] | null;
  createdAt: Date;
  updatedAt: Date;
}