# lib/api-zod/src/generated/types/updateArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for The Front Porch Bulletin
 * OpenAPI spec version: 0.1.0
 */
import type { Photo } from "./photo";

export interface UpdateArticleBody {
  title?: string;
  subtitle?: string | null;
  content?: string;
  author?: string;
  category?: string;
  featured?: boolean;
  photoUrl?: string | null;
  photos?: Photo[] | null;
  publishedAt?: Date;
}